+++ connectors/kafka/kafka-connector/src/main/java/org/platformlambda/kafka/KafkaSetup.java
@@ -38,6 +38,7 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.TimeoutException;
 
@@ -47,13 +48,16 @@ public class KafkaSetup implements CloudSetup {
 
     public static final String MANAGER = "kafka.manager";
     public static final String PRESENCE_MONITOR = "presence.monitor";
+    public static final String KAFKA_PROPERTIES_PROVIDER = "kafka.properties.provider";
     private static final String CLOUD_CHECK = "cloud.connector.health";
     private static final String ORIGIN = "origin";
     private static final String BROKER_URL = "bootstrap.servers";
     private static final String TYPE = "type";
     private static final String CREATE_TOPIC = "create_topic";
+    private static final String GET = "get";
 
     private static String displayUrl = "unknown";
+    private static boolean pingKafka = true;
     private boolean isServiceMonitor;
 
 
@@ -66,50 +70,96 @@ public class KafkaSetup implements CloudSetup {
         return displayUrl;
     }
 
-    @Override
-    public void initialize() {
-        try {
+    @SuppressWarnings("unchecked")
+    public static Properties getKafkaProperties() {
+        Platform platform = Platform.getInstance();
+        PostOffice po = PostOffice.getInstance();
+        Properties properties = new Properties();
+        if (platform.hasRoute(KAFKA_PROPERTIES_PROVIDER)) {
+            /*
+             * some Kafka installations regularly change kafka credentials.
+             * To handle this case, the wrapper for kafka connector must implement service
+             * for KAFKA_PROPERTIES_PROVIDER.
+             */
+            try {
+                EventEnvelope res = po.request(KAFKA_PROPERTIES_PROVIDER, 5000, new Kv(TYPE, GET));
+                if (res.getBody() instanceof Map) {
+                    Map<String, Object> map = (Map<String, Object>) res.getBody();
+                    for (String key: map.keySet()) {
+                        String value = map.get(key).toString();
+                        properties.setProperty(key, value);
+                    }
+                } else {
+                    throw new IOException("Expect: map in response, Actual: "+
+                            (res.getBody() == null? "null" : res.getBody().getClass().getSimpleName()));
+                }
+            } catch (IOException | TimeoutException | AppException e) {
+                log.error("Unable to obtain kafka.properties from {}  - {}", KAFKA_PROPERTIES_PROVIDER, e.getMessage());
+                System.exit(-1);
+            }
+        } else {
+            /*
+             * Retrieve kafka.properties from the local file system.
+             * This assumes the configuration is created by CI/CD or a wrapper for kafka connector.
+             */
             AppConfigReader reader = AppConfigReader.getInstance();
-            // to find kafka.properties, try file first, then classpath
             String pathList = reader.getProperty("kafka.client.properties",
-                                          "file:/tmp/config/kafka.properties,classpath:/kafka.properties");
+                    "file:/tmp/config/kafka.properties,classpath:/kafka.properties");
             ConfigReader config = SetupUtil.getConfig(pathList);
             if (config == null) {
                 log.error("Unable to find kafka properties from {}", pathList);
                 System.exit(-1);
             }
-            String brokerUrls = config.getProperty(BROKER_URL);
-            List<String> brokers = Utility.getInstance().split(brokerUrls, ",");
-            if (brokers.isEmpty()) {
-                log.error("Unable to setup kafka from {} - missing {}", pathList, BROKER_URL);
-                System.exit(-1);
-            }
-            if (brokers.size() > 1) {
-                Collections.sort(brokers);
-            }
-            // try 2 times to check if kafka cluster is available
-            if (!kafkaReachable(brokers, 2)) {
-                log.error("Kafka cluster is not reachable {}", brokers);
-                System.exit(-1);
-            }
-            // use the first broker URL as the display URL in the info endpoint
-            displayUrl = brokers.get(0);
-            // build kafka base properties
-            Properties baseProp = new Properties();
             for (String k: config.getMap().keySet()) {
-                baseProp.put(k, config.getProperty(k));
+                properties.setProperty(k, config.getProperty(k));
             }
+        }
+        String brokerUrls = properties.getProperty(BROKER_URL);
+        List<String> brokers = Utility.getInstance().split(brokerUrls, ",");
+        if (brokers.isEmpty()) {
+            log.error("Unable to setup kafka - missing {}", BROKER_URL);
+            System.exit(-1);
+        }
+        if (brokers.size() > 1) {
+            Collections.sort(brokers);
+        }
+        if (pingKafka) {
+            /*
+             * Ping Kafka cluster when the application starts up.
+             * This assumes the broker list is constant over the life time of the application.
+             */
+            pingKafka = false;
+            try {
+                // try 2 times to check if kafka cluster is available
+                if (!kafkaReachable(brokers, 2)) {
+                    throw new IOException("Unreachable");
+                }
+            } catch (IOException e) {
+                log.error("Kafka cluster failure {} - {}", brokers, e.getMessage());
+                System.exit(-1);
+            }
+        }
+        // use the first broker URL as the display URL in the info endpoint
+        displayUrl = brokers.get(0);
+        return properties;
+    }
+
+    @Override
+    public void initialize() {
+        try {
+            // ensure Kafka properties can be retrieved
+            Properties properties = getKafkaProperties();
             /*
              * Enable pub/sub feature in case the user application wants to do pub/sub manually
              * (it must be enabled before starting cloud connector
              *  because pub/sub will check the dependency of the connector)
              */
-            PubSub.getInstance().enableFeature(new KafkaPubSub(baseProp));
+            PubSub.getInstance().enableFeature(new KafkaPubSub());
             // setup producer
             Platform platform = Platform.getInstance();
             log.info("Starting kafka producer module");
-            platform.registerPrivate(PostOffice.CLOUD_CONNECTOR, new EventProducer(baseProp), 1);
-            platform.registerPrivate(MANAGER, new TopicManager(baseProp), 1);
+            platform.registerPrivate(PostOffice.CLOUD_CONNECTOR, new EventProducer(), 1);
+            platform.registerPrivate(MANAGER, new TopicManager(), 1);
             String origin = platform.getOrigin();
             log.info("Starting kafka consumer module");
             String namespace = Platform.getInstance().getNamespace();
@@ -124,7 +174,7 @@ public class KafkaSetup implements CloudSetup {
                 }
             }
             // setup consumer and connect to Kafka
-            EventConsumer consumer = new EventConsumer(baseProp, topic);
+            EventConsumer consumer = new EventConsumer(properties, topic);
             consumer.start();
             Runtime.getRuntime().addShutdownHook(new Thread(consumer::shutdown));
             // enable service discovery
@@ -132,7 +182,6 @@ public class KafkaSetup implements CloudSetup {
             platform.registerPrivate(ServiceDiscovery.SERVICE_QUERY, new ServiceQuery(), 10);
             platform.registerPrivate(CLOUD_CHECK, new KafkaHealthCheck(), 2);
             platform.startCloudServices();
-
             if (!isServiceMonitor) {
                 AppAlive alive = new AppAlive();
                 alive.start();
@@ -144,7 +193,7 @@ public class KafkaSetup implements CloudSetup {
         }
     }
 
-    private boolean kafkaReachable(List<String> brokers, int tries) throws IOException {
+    private static boolean kafkaReachable(List<String> brokers, int tries) throws IOException {
         for (String dest: brokers) {
             if (dest.contains(":")) {
                 Utility util = Utility.getInstance();